fastapi-projects/17-saathi-agent/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
websockets==14.1
jinja2==3.1.4
pydantic==2.10.4
